+++ apps/server/src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod account;
+pub mod gifs;
 pub mod images;
 pub mod pools;
 pub mod random;